🔨 refactor(model-runtime): extract model processing to shared utility
- Create pullModelsPostProcess utility with processModelListWithImageModels
- Refactor openaiCompatibleFactory and RouterRuntime to use shared logic
- Reduce code duplication for model type processing and image model generation
